gmovies/pyproject.toml
[tool.poetry]
name = "gmovies"
version = "0.1.0"
description = "Select movies to watch with your friends based on streaming availability."
authors = ["Gewad <thegewad@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
discord = "^2.2.2"
python-dotenv = "^1.0.0"
zodb = "^5.8.0"
types-requests = "^2.28.11.16"

[tool.poetry.group.dev.dependencies]
nox-poetry = "^1.0.2"
flake8 = "^6.0.0"
black = "^23.1.0"
bandit = "^1.7.5"
mypy = "^1.1.1"
nox = "^2022.11.21"
pytest = "^7.2.2"
pyupgrade = "^3.3.1"
isort = "^5.12.0"
flake8-pyproject = "^1.2.3"
flake8-bugbear = "^23.3.23"
flake8-black = "^0.3.6"
flake8-isort = "^6.0.0"
autoflake = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 100
